refactor: project_detail 구조 개편
- widgets/read_only/, widgets/editable/을 만들어 관리
- 기존 코드를 분리하여 파일단위로 역할을 나눔
