+++ app/lib/src/pages/project_page/project_detail_page.dart
@@ -3,14 +3,14 @@ import 'package:provider/provider.dart';
 
 import 'package:talk_pilot/src/models/project_model.dart';
 import 'package:talk_pilot/src/provider/user_provider.dart';
-
 import 'package:talk_pilot/src/components/loading_indicator.dart';
+
 import 'package:talk_pilot/src/services/database/project_service.dart';
 import 'package:talk_pilot/src/services/database/project_stream_service.dart';
 
-import 'package:talk_pilot/src/pages/project_page/widgets/text_editor.dart';
-import 'package:talk_pilot/src/pages/project_page/widgets/script_upload.dart';
 import 'package:talk_pilot/src/pages/project_page/widgets/project_info_card.dart';
+import 'package:talk_pilot/src/pages/project_page/widgets/script_upload_button.dart';
+import 'package:talk_pilot/src/pages/project_page/widgets/editable/editable_text_editor.dart';
 
 class ProjectDetailPage extends StatefulWidget {
   final String projectId;
@@ -25,22 +25,17 @@ class _ProjectDetailPageState extends State<ProjectDetailPage> {
   bool isLoading = false;
 
   ProjectRole getUserRole(ProjectModel project, String? uid) {
-    final role = project.participants[uid] ?? 'MEMBER';
-    switch (role.toLowerCase()) {
-      case 'owner':
-        return ProjectRole.owner;
-      case 'editor':
-        return ProjectRole.editor;
-      case 'member':
-      default:
-        return ProjectRole.member;
-    }
+    final role = project.participants[uid] ?? 'member';
+    return switch (role.toLowerCase()) {
+      'owner' => ProjectRole.owner,
+      'editor' => ProjectRole.editor,
+      _ => ProjectRole.member,
+    };
   }
 
   @override
   Widget build(BuildContext context) {
-    final currentUser = context.read<UserProvider>().currentUser;
-    final currentUid = currentUser?.uid;
+    final currentUid = context.read<UserProvider>().currentUser?.uid;
 
     return StreamBuilder<ProjectModel>(
       stream: _projectService.streamProject(widget.projectId),
@@ -55,9 +50,8 @@ class _ProjectDetailPageState extends State<ProjectDetailPage> {
         }
 
         final project = snapshot.data!;
-        final role = getUserRole(project, currentUid);
         final isEditable =
-            role == ProjectRole.owner || role == ProjectRole.editor;
+            getUserRole(project, currentUid) != ProjectRole.member;
 
         return Scaffold(
           appBar: AppBar(
@@ -70,7 +64,7 @@ class _ProjectDetailPageState extends State<ProjectDetailPage> {
             iconTheme: const IconThemeData(color: Colors.white),
             actions: [
               if (isEditable)
-                ScriptUpload(
+                ScriptUploadButton(
                   context: context,
                   isLoading: isLoading,
                   projectId: widget.projectId,
@@ -90,42 +84,40 @@ class _ProjectDetailPageState extends State<ProjectDetailPage> {
               const SizedBox(height: 12),
               ProjectInfoCard(project: project, editable: isEditable),
               const SizedBox(height: 32),
-              TextEditor(
-                projectId: project.id,
-                field: ProjectField.title,
-                label: '제목',
-                value: project.title,
-                maxLength: 100,
-                editable: isEditable,
-              ),
-              const SizedBox(height: 16),
-              TextEditor(
-                projectId: project.id,
-                field: ProjectField.description,
-                label: '설명',
-                value: project.description,
-                maxLength: 300,
-                editable: isEditable,
-              ),
-              const SizedBox(height: 16),
-              TextEditor(
-                projectId: project.id,
-                field: ProjectField.memo,
-                label: '메모',
-                value: project.memo ?? '',
-                maxLength: 1000,
-                editable: isEditable,
-              ),
-              const SizedBox(height: 16),
-              TextEditor(
-                projectId: project.id,
-                field: ProjectField.script,
-                label: '대본',
-                value: project.script ?? '',
-                maxLength: 3000,
-                editable: isEditable,
-              ),
-              const SizedBox(height: 32),
+              ...[
+                EditableTextEditor(
+                  projectId: project.id,
+                  field: ProjectField.title,
+                  label: '제목',
+                  value: project.title,
+                  maxLength: 100,
+                  editable: isEditable,
+                ),
+                EditableTextEditor(
+                  projectId: project.id,
+                  field: ProjectField.description,
+                  label: '설명',
+                  value: project.description,
+                  maxLength: 300,
+                  editable: isEditable,
+                ),
+                EditableTextEditor(
+                  projectId: project.id,
+                  field: ProjectField.memo,
+                  label: '메모',
+                  value: project.memo ?? '',
+                  maxLength: 1000,
+                  editable: isEditable,
+                ),
+                EditableTextEditor(
+                  projectId: project.id,
+                  field: ProjectField.script,
+                  label: '대본',
+                  value: project.script ?? '',
+                  maxLength: 3000,
+                  editable: isEditable,
+                ),
+              ].expand((widget) => [widget, const SizedBox(height: 16)]),
             ],
           ),
         );
